Modularize Streamlit UI for maintainability
Split the monolithic 2,109-line app.py into 12 focused modules
organized by responsibility. Use _pages/ (underscore prefix) to
prevent Streamlit's auto-discovery from creating phantom sidebar
entries. No behavior changes -- pure code reorganization.
