+++ src/ui/_pages/__init__.py
@@ -1,0 +1,1 @@
+"""ChemGraph UI page modules."""
