+++ src/ui/agent_manager.py
@@ -1,0 +1,42 @@
+"""Agent lifecycle management for the ChemGraph Streamlit UI."""
+
+from typing import Optional
+
+import streamlit as st
+
+
+@st.cache_resource
+def initialize_agent(
+    model_name: str,
+    workflow_type: str,
+    structured_output: bool,
+    return_option: str,
+    generate_report: bool,
+    recursion_limit: int,
+    base_url: Optional[str],
+    argo_user: Optional[str],
+):
+    """Create (or reuse) a cached :class:`ChemGraph` agent instance."""
+    try:
+        from chemgraph.agent.llm_agent import ChemGraph
+
+        return ChemGraph(
+            model_name=model_name,
+            workflow_type=workflow_type,
+            base_url=base_url,
+            argo_user=argo_user,
+            structured_output=structured_output,
+            generate_report=generate_report,
+            return_option=return_option,
+            recursion_limit=recursion_limit,
+        )
+    except Exception as exc:
+        st.error(f"Failed to initialize agent: {exc}")
+        return None
+
+
+def run_async_callable(fn):
+    """Run an async callable and return its result in a sync context."""
+    from chemgraph.utils.async_utils import run_async_callable as _impl
+
+    return _impl(fn)
