+++ src/ui/app.py
@@ -1,39 +1,31 @@
-import ast
-from datetime import datetime, timezone, timedelta
-import html as html_mod
-import json
-import os
-import platform
+"""ChemGraph Streamlit application entry point.
+
+Run with:  ``streamlit run src/ui/app.py``
+
+This thin module handles page configuration (which **must** be the first
+Streamlit call), sidebar navigation, and page dispatch.  All page content
+lives in :mod:`ui.pages`.
+"""
+
+import sys
 from pathlib import Path
-import re
-import socket
-import subprocess
-from typing import Optional, Dict, Any
-from urllib.error import HTTPError, URLError
-from urllib.parse import urlparse
-from urllib.request import Request, urlopen
-from uuid import uuid4
 
-from ase.data import chemical_symbols
-from ase.io import read as ase_read
-import numpy as np
-import pandas as pd
+# Ensure the parent of ui/ (i.e. src/) is on sys.path so that
+# "from ui.xxx import ..." works when run as a standalone script.
+_SRC_DIR = str(Path(__file__).resolve().parent.parent)
+if _SRC_DIR not in sys.path:
+    sys.path.insert(0, _SRC_DIR)
+
 import streamlit as st
-import toml
 
-import chemgraph as chemgraph_pkg
 from chemgraph import __version__ as chemgraph_version
-from chemgraph.tools.ase_tools import create_ase_atoms, create_xyz_string
-from chemgraph.models.supported_models import (
-    supported_argo_models,
-)
-from chemgraph.utils.config_utils import (
-    get_argo_user_from_nested_config,
-    get_base_url_for_model_from_nested_config,
-    get_model_options_for_nested_config,
-)
 
-# Page configuration -- MUST be first Streamlit call
+from ui.system_info import render_sidebar_host_and_build_info
+from ui.visualization import warn_stmol_unavailable
+
+# ---------------------------------------------------------------------------
+# Page configuration -- MUST be the first Streamlit call
+# ---------------------------------------------------------------------------
 app_version = (
     chemgraph_version
     if isinstance(chemgraph_version, str) and chemgraph_version != "unknown"
@@ -42,2068 +34,42 @@ app_version = (
 
 st.set_page_config(
     page_title="ChemGraph",
-    page_icon="🧪",
+    page_icon="\U0001f9ea",
     layout="wide",
     initial_sidebar_state="expanded",
 )
 
-WORKFLOW_ALIASES = {
-    "python_repl": "python_relp",
-    "graspa_agent": "graspa",
-}
-WORKFLOW_OPTIONS = [
-    "single_agent",
-    "multi_agent",
-    "python_relp",
-    "graspa",
-    "mock_agent",
-]
+# One-time stmol availability warning
+warn_stmol_unavailable()
 
-
-def normalize_workflow_name(value: str) -> str:
-    if not value:
-        return value
-    return WORKFLOW_ALIASES.get(value, value)
-
-
-def resolve_output_path(path: str) -> str:
-    """Resolve output paths relative to CHEMGRAPH_LOG_DIR when set."""
-    if not path:
-        return path
-    if os.path.isabs(path):
-        return path
-    log_dir = os.environ.get("CHEMGRAPH_LOG_DIR")
-    if log_dir:
-        return os.path.join(log_dir, path)
-    return path
-
-
-def get_base_url_for_model(model_name: str, config: Dict[str, Any]) -> Optional[str]:
-    return get_base_url_for_model_from_nested_config(model_name, config)
-
-
-def get_model_options(config: Dict[str, Any]) -> list:
-    return get_model_options_for_nested_config(config)
-
-
-def run_async_callable(fn):
-    """Run an async callable and return its result in sync context."""
-    from chemgraph.utils.async_utils import run_async_callable as _impl
-
-    return _impl(fn)
-
-
-def _run_command(cmd: list[str], cwd: Optional[Path] = None, timeout: int = 2) -> str:
-    """Run a shell command and return stripped stdout; return empty string on failure."""
-    try:
-        completed = subprocess.run(
-            cmd,
-            capture_output=True,
-            text=True,
-            timeout=timeout,
-            check=True,
-            cwd=str(cwd) if cwd else None,
-        )
-    except Exception:
-        return ""
-    return completed.stdout.strip()
-
-
-def _find_repo_root(start: Path) -> Optional[Path]:
-    """Find git repo root by walking up parents from a starting path."""
-    start = start.resolve()
-    candidates = [start] + list(start.parents)
-    for candidate in candidates:
-        if (candidate / ".git").exists():
-            return candidate
-    return None
-
-
-def _format_bytes(num_bytes: int) -> str:
-    if num_bytes <= 0:
-        return "Unknown"
-    units = ["B", "KB", "MB", "GB", "TB", "PB"]
-    size = float(num_bytes)
-    for unit in units:
-        if size < 1024.0 or unit == units[-1]:
-            return f"{size:.1f} {unit}"
-        size /= 1024.0
-    return "Unknown"
-
-
-def _get_total_memory_bytes() -> int:
-    """Return total system memory in bytes when available."""
-    try:
-        page_size = os.sysconf("SC_PAGE_SIZE")
-        phys_pages = os.sysconf("SC_PHYS_PAGES")
-        total = int(page_size) * int(phys_pages)
-        if total > 0:
-            return total
-    except Exception:
-        pass
-
-    meminfo = Path("/proc/meminfo")
-    if meminfo.exists():
-        try:
-            for line in meminfo.read_text().splitlines():
-                if line.startswith("MemTotal:"):
-                    kb = int(line.split()[1])
-                    return kb * 1024
-        except Exception:
-            return 0
-    return 0
-
-
-def _get_cpu_model() -> str:
-    """Try to get a human-readable CPU model name."""
-    cpuinfo = Path("/proc/cpuinfo")
-    if cpuinfo.exists():
-        try:
-            for line in cpuinfo.read_text().splitlines():
-                if line.lower().startswith("model name"):
-                    parts = line.split(":", 1)
-                    if len(parts) == 2:
-                        return parts[1].strip()
-        except Exception:
-            pass
-
-    cpu_name = platform.processor().strip()
-    if cpu_name:
-        return cpu_name
-    return platform.machine()
-
-
-def _get_gpu_summary() -> str:
-    """Return GPU summary from nvidia-smi when available."""
-    output = _run_command(
-        [
-            "nvidia-smi",
-            "--query-gpu=name,memory.total",
-            "--format=csv,noheader,nounits",
-        ]
-    )
-    if not output:
-        return "No GPU detected"
-
-    entries = []
-    for line in output.splitlines():
-        parts = [part.strip() for part in line.split(",")]
-        if len(parts) >= 2:
-            name, mem_mib = parts[0], parts[1]
-            entries.append(f"{name} ({mem_mib} MiB)")
-        elif parts:
-            entries.append(parts[0])
-    return "; ".join(entries) if entries else "No GPU detected"
-
-
-@st.cache_data(ttl=60)
-def get_host_info() -> Dict[str, str]:
-    """Collect host metadata for sidebar display."""
-    return {
-        "hostname": socket.gethostname(),
-        "platform": f"{platform.system()} {platform.release()}",
-        "cpu_model": _get_cpu_model(),
-        "cpu_cores": str(os.cpu_count() or "Unknown"),
-        "memory_total": _format_bytes(_get_total_memory_bytes()),
-        "gpu": _get_gpu_summary(),
-    }
-
-
-@st.cache_data(ttl=60)
-def get_build_info() -> Dict[str, str]:
-    """Collect app and repository metadata for sidebar display."""
-    app_file = Path(__file__).resolve()
-    chemgraph_file = Path(chemgraph_pkg.__file__).resolve()
-    repo_root = _find_repo_root(app_file) or _find_repo_root(chemgraph_file)
-
-    commit = "Unknown"
-    commit_date = "Unknown"
-    branch = "Unknown"
-
-    if repo_root:
-        commit = _run_command(["git", "rev-parse", "--short", "HEAD"], cwd=repo_root) or "Unknown"
-        commit_date = (
-            _run_command(["git", "show", "-s", "--format=%cd", "--date=iso", "HEAD"], cwd=repo_root)
-            or "Unknown"
-        )
-        branch = _run_command(["git", "rev-parse", "--abbrev-ref", "HEAD"], cwd=repo_root) or "Unknown"
-
-    return {
-        "chemgraph_version": str(chemgraph_version),
-        "commit": commit,
-        "commit_date": commit_date,
-        "branch": branch,
-        "chemgraph_file": str(chemgraph_file),
-    }
-
-
-def render_sidebar_host_and_build_info():
-    """Render host and build metadata blocks in the left sidebar."""
-    host_info = get_host_info()
-    build_info = get_build_info()
-
-    with st.sidebar.expander("🖥️ Host Info", expanded=False):
-        st.markdown(f"**Hostname:** `{host_info['hostname']}`")
-        st.markdown(f"**OS:** `{host_info['platform']}`")
-        st.markdown(f"**CPU:** `{host_info['cpu_model']}`")
-        st.markdown(f"**CPU Cores:** `{host_info['cpu_cores']}`")
-        st.markdown(f"**Memory:** `{host_info['memory_total']}`")
-        st.markdown(f"**GPU:** `{host_info['gpu']}`")
-
-    with st.sidebar.expander("📦 Build Info", expanded=False):
-        st.markdown(f"**ChemGraph Version:** `{build_info['chemgraph_version']}`")
-        st.markdown(f"**Branch:** `{build_info['branch']}`")
-        st.markdown(f"**Commit:** `{build_info['commit']}`")
-        st.markdown(f"**Commit Date:** `{build_info['commit_date']}`")
-        st.markdown(f"**ChemGraph File:** `{build_info['chemgraph_file']}`")
-
-
-def _is_local_address(hostname: str) -> bool:
-    host = (hostname or "").strip().lower()
-    return host in {"localhost", "127.0.0.1", "0.0.0.0", "::1"}
-
-
-@st.cache_data(ttl=10)
-def check_local_model_endpoint(base_url: Optional[str]) -> Dict[str, Any]:
-    """Quick reachability check for local OpenAI-compatible endpoints."""
-    if not base_url:
-        return {"ok": True, "message": "No base URL configured."}
-
-    parsed = urlparse(base_url)
-    if not _is_local_address(parsed.hostname or ""):
-        return {"ok": True, "message": "Skipping non-local endpoint probe."}
-
-    probe = base_url.rstrip("/") + "/models"
-    req = Request(probe, method="GET")
-
-    try:
-        with urlopen(req, timeout=2) as response:
-            code = getattr(response, "status", 200)
-            return {"ok": True, "message": f"Reachable (HTTP {code})."}
-    except HTTPError as e:
-        # HTTP error still means service/socket is reachable.
-        return {"ok": True, "message": f"Reachable (HTTP {e.code})."}
-    except URLError as e:
-        reason = getattr(e, "reason", e)
-        return {"ok": False, "message": f"Unreachable: {reason}"}
-    except Exception as e:
-        return {"ok": False, "message": f"Unreachable: {e}"}
-
-
-# Configuration management
-try:
-    from .config import load_config, save_config, get_default_config
-except ImportError:
-    # Handle case when running as script (not as package)
-    import sys
-
-    # Get current directory - handle both package and script execution
-    if "__file__" in globals():
-        current_dir = os.path.dirname(os.path.abspath(__file__))
-    else:
-        current_dir = os.getcwd()
-
-    if current_dir not in sys.path:
-        sys.path.insert(0, current_dir)
-
-    try:
-        from config import load_config, save_config, get_default_config
-    except ImportError:
-        # Fallback: assume we're in the project root
-        config_dir = os.path.join(os.getcwd(), "src", "ui")
-        if config_dir not in sys.path:
-            sys.path.insert(0, config_dir)
-        from config import load_config, save_config, get_default_config
-
-
-# -----------------------------------------------------------------------------
-# Optional 3-D viewer - stmol + py3Dmol
-# -----------------------------------------------------------------------------
-try:
-    import stmol
-
-    STMOL_AVAILABLE = True
-except ImportError:
-    STMOL_AVAILABLE = False
-    st.warning("⚠️ **stmol** not available – falling back to text/table view.")
-    st.info("To enable 3D visualization, install with: `pip install stmol`")
-
-# -----------------------------------------------------------------------------
-# Page Navigation
-# -----------------------------------------------------------------------------
-st.sidebar.title("🧪 ChemGraph")
+# ---------------------------------------------------------------------------
+# Sidebar navigation
+# ---------------------------------------------------------------------------
+st.sidebar.title("\U0001f9ea ChemGraph")
 page = st.sidebar.radio(
     "Navigate",
-    ["🏠 Main Interface", "⚙️ Configuration", "📖 About ChemGraph"],
+    ["\U0001f3e0 Main Interface", "\u2699\ufe0f Configuration", "\U0001f4d6 About ChemGraph"],
     index=0,
     key="page_navigation",
 )
 render_sidebar_host_and_build_info()
 
-# -----------------------------------------------------------------------------
-# About Page
-# -----------------------------------------------------------------------------
-if page == "📖 About ChemGraph":
-    st.title("📖 About ChemGraph")
+# ---------------------------------------------------------------------------
+# Page dispatch
+# ---------------------------------------------------------------------------
+if page == "\U0001f4d6 About ChemGraph":
+    from ui._pages import about
 
-    st.markdown(
-        """
-    ## AI Agents for Computational Chemistry
-    
-    ChemGraph is an **agentic framework** for computational chemistry and materials science workflows. 
-    It enables researchers to perform complex computational chemistry tasks using natural language queries 
-    powered by large language models (LLMs) and specialized AI agents.
-    
-    ### 🔬 Key Features
-    
-    - **Multi-Agent Workflows**: Coordinate multiple AI agents for complex computational tasks
-    - **Natural Language Interface**: Interact with computational chemistry tools using plain English
-    - **Molecular Visualization**: 3D interactive molecular structure visualization
-    - **Multiple Calculators**: Support for various quantum chemistry packages (ORCA, Psi4, MACE, etc.)
-    - **Report Generation**: Automated generation of computational chemistry reports
-    - **Flexible Backends**: Support for various LLM providers (OpenAI, Anthropic, local models)
-    
-    ### 📚 Resources
-    
-    - **GitHub**: [https://github.com/argonne-lcf/ChemGraph](https://github.com/argonne-lcf/ChemGraph)
-    - **Documentation**: [https://argonne-lcf.github.io/ChemGraph/](https://argonne-lcf.github.io/ChemGraph/)
-    
-    ### 🏛️ Developed at Argonne National Laboratory
-    
-    ChemGraph is developed at **Argonne National Laboratory** as part of advancing 
-    computational chemistry and materials science research through AI-driven automation.
-    
-    ### 📄 License
-    
-    This project is licensed under the **Apache License 2.0** - see the 
-    [LICENSE](https://github.com/argonne-lcf/ChemGraph/blob/main/LICENSE) file for details.
-    
-    ### 🙏 Citation
-    
-    If you use ChemGraph in your research, please cite our [work](https://doi.org/10.1038/s42004-025-01776-9):
-    
-    ```bibtex
-    @article{pham_chemgraph_2026,
-    title = {{ChemGraph} as an agentic framework for computational chemistry workflows},
-    url = {https://doi.org/10.1038/s42004-025-01776-9},
-    doi = {10.1038/s42004-025-01776-9},
-    author = {Pham, Thang D. and Tanikanti, Aditya and Ke\c{c}eli, Murat},
-    date = {2026-01-08},
-    author={Pham, Thang D and Tanikanti, Aditya and Ke{\c{c}}eli, Murat},
-    journal={Communications Chemistry},
-    year={2026},
-    publisher={Nature Publishing Group UK London}
-    }
-    ```
-
-    ### 🙌 Acknowledgments
-
-    This research used resources of the Argonne Leadership Computing Facility, a U.S.
-    Department of Energy (DOE) Office of Science user facility at Argonne National
-    Laboratory and is based on research supported by the U.S. DOE Office of Science-
-    Advanced Scientific Computing Research Program, under Contract No. DE-AC02-
-    06CH11357. Our work leverages ALCF Inference Endpoints, which provide a robust API
-    for LLM inference on ALCF HPC clusters via Globus Compute. We are thankful to Serkan
-    Altuntas for his contributions to the user interface of ChemGraph and for insightful
-    discussions on AIOps.
-    
-    ---
-    
-    ### 🚀 Get Started
-    
-    Ready to use ChemGraph? Switch to the **🏠 Main Interface** using the navigation menu on the left 
-    to start running computational chemistry workflows with AI agents!
-    """
-    )
-
-    # Stop execution here for About page
+    about.render()
     st.stop()
 
-# -----------------------------------------------------------------------------
-# Configuration Page
-# -----------------------------------------------------------------------------
-elif page == "⚙️ Configuration":
-    st.title("⚙️ Configuration")
-    st.markdown(
-        """
-    Edit and manage your ChemGraph configuration settings. Changes are saved to `config.toml`.
-    """
-    )
+elif page == "\u2699\ufe0f Configuration":
+    from ui._pages import configuration
 
-    # Initialize session state for config
-    if "config" not in st.session_state:
-        st.session_state.config = load_config()
-
-    config = st.session_state.config
-
-    # Configuration tabs
-    tab1, tab2, tab3 = st.tabs(
-        ["🔧 General Settings", "🔗 API Settings", "📝 Raw TOML"]
-    )
-
-    with tab1:
-        st.subheader("General Settings")
-
-        col1, col2 = st.columns(2)
-
-        with col1:
-            st.write("**Model & Workflow**")
-            model_options = get_model_options(config)
-            config["general"]["model"] = st.selectbox(
-                "Model",
-                model_options,
-                index=(
-                    model_options.index(config["general"]["model"])
-                    if config["general"]["model"] in model_options
-                    else 0
-                ),
-                key="config_model",
-            )
-            config_custom_model = st.text_input(
-                "Custom model ID (optional)",
-                value="",
-                key="config_custom_model",
-                help="Enter any provider/model identifier not listed above.",
-            ).strip()
-            if config_custom_model:
-                config["general"]["model"] = config_custom_model
-
-            config["general"]["workflow"] = normalize_workflow_name(
-                config["general"]["workflow"]
-            )
-            config["general"]["workflow"] = st.selectbox(
-                "Workflow",
-                WORKFLOW_OPTIONS,
-                index=(
-                    WORKFLOW_OPTIONS.index(config["general"]["workflow"])
-                    if config["general"]["workflow"] in WORKFLOW_OPTIONS
-                    else 0
-                ),
-                key="config_workflow",
-            )
-
-            config["general"]["output"] = st.selectbox(
-                "Output Format",
-                ["state", "last_message"],
-                index=(
-                    ["state", "last_message"].index(config["general"]["output"])
-                    if config["general"]["output"] in ["state", "last_message"]
-                    else 0
-                ),
-                key="config_output",
-            )
-
-            config["general"]["structured"] = st.checkbox(
-                "Structured Output",
-                value=config["general"]["structured"],
-                key="config_structured",
-            )
-
-            config["general"]["report"] = st.checkbox(
-                "Generate Report",
-                value=config["general"]["report"],
-                key="config_report",
-            )
-
-            config["general"]["verbose"] = st.checkbox(
-                "Verbose Output",
-                value=config["general"]["verbose"],
-                key="config_verbose",
-            )
-
-        with col2:
-            st.write("**Execution Settings**")
-            config["general"]["thread"] = st.number_input(
-                "Thread ID",
-                min_value=1,
-                max_value=1000,
-                value=config["general"]["thread"],
-                key="config_thread",
-            )
-
-            config["general"]["recursion_limit"] = st.number_input(
-                "Recursion Limit",
-                min_value=1,
-                max_value=100,
-                value=config["general"]["recursion_limit"],
-                key="config_recursion",
-            )
-
-        st.subheader("Chemistry Settings")
-
-        col3, col4 = st.columns(2)
-
-        with col3:
-            st.write("**Optimization**")
-            config["chemistry"]["optimization"]["method"] = st.selectbox(
-                "Method",
-                ["BFGS", "L-BFGS-B", "CG", "Newton-CG"],
-                index=(
-                    ["BFGS", "L-BFGS-B", "CG", "Newton-CG"].index(
-                        config["chemistry"]["optimization"]["method"]
-                    )
-                    if config["chemistry"]["optimization"]["method"]
-                    in ["BFGS", "L-BFGS-B", "CG", "Newton-CG"]
-                    else 0
-                ),
-                key="config_opt_method",
-            )
-
-            config["chemistry"]["optimization"]["fmax"] = st.number_input(
-                "Force Max (eV/Å)",
-                min_value=0.001,
-                max_value=1.0,
-                value=config["chemistry"]["optimization"]["fmax"],
-                format="%.3f",
-                key="config_fmax",
-            )
-
-            config["chemistry"]["optimization"]["steps"] = st.number_input(
-                "Max Steps",
-                min_value=1,
-                max_value=1000,
-                value=config["chemistry"]["optimization"]["steps"],
-                key="config_steps",
-            )
-
-        with col4:
-            st.write("**Calculators**")
-            calc_options = [
-                "mace_mp",
-                "mace_off",
-                "mace_anicc",
-                "fairchem",
-                "aimnet2",
-                "emt",
-                "tblite",
-                "orca",
-                "nwchem",
-            ]
-            config["chemistry"]["calculators"]["default"] = st.selectbox(
-                "Default Calculator",
-                calc_options,
-                index=(
-                    calc_options.index(config["chemistry"]["calculators"]["default"])
-                    if config["chemistry"]["calculators"]["default"] in calc_options
-                    else 0
-                ),
-                key="config_calc_default",
-            )
-
-            config["chemistry"]["calculators"]["fallback"] = st.selectbox(
-                "Fallback Calculator",
-                calc_options,
-                index=(
-                    calc_options.index(config["chemistry"]["calculators"]["fallback"])
-                    if config["chemistry"]["calculators"]["fallback"] in calc_options
-                    else 1
-                ),
-                key="config_calc_fallback",
-            )
-
-    with tab2:
-        st.subheader("API Settings")
-
-        st.markdown("**API Keys (Session Only)**")
-        st.caption(
-            "Keys entered here are applied to this Streamlit session via environment variables and are not saved to config.toml."
-        )
-
-        key_col1, key_col2 = st.columns(2)
-        with key_col1:
-            openai_api_key = st.text_input(
-                "OpenAI API Key",
-                value=st.session_state.get("ui_openai_api_key", ""),
-                type="password",
-                key="ui_openai_api_key_input",
-            )
-            anthropic_api_key = st.text_input(
-                "Anthropic API Key",
-                value=st.session_state.get("ui_anthropic_api_key", ""),
-                type="password",
-                key="ui_anthropic_api_key_input",
-            )
-        with key_col2:
-            gemini_api_key = st.text_input(
-                "Gemini API Key",
-                value=st.session_state.get("ui_gemini_api_key", ""),
-                type="password",
-                key="ui_gemini_api_key_input",
-            )
-            groq_api_key = st.text_input(
-                "Groq API Key",
-                value=st.session_state.get("ui_groq_api_key", ""),
-                type="password",
-                key="ui_groq_api_key_input",
-            )
-
-        key_env_map = {
-            "OPENAI_API_KEY": openai_api_key,
-            "ANTHROPIC_API_KEY": anthropic_api_key,
-            "GEMINI_API_KEY": gemini_api_key,
-            "GROQ_API_KEY": groq_api_key,
-        }
-
-        action_col1, action_col2 = st.columns(2)
-        with action_col1:
-            if st.button("Apply API Keys", key="apply_api_keys"):
-                applied = []
-                for env_name, key_value in key_env_map.items():
-                    clean_value = key_value.strip()
-                    if clean_value:
-                        os.environ[env_name] = clean_value
-                        st.session_state[f"ui_{env_name.lower()}"] = clean_value
-                        applied.append(env_name)
-                if applied:
-                    st.success(f"✅ Applied keys for: {', '.join(applied)}")
-                else:
-                    st.info("No API keys entered.")
-        with action_col2:
-            if st.button("Clear Session API Keys", key="clear_api_keys"):
-                for env_name in key_env_map:
-                    os.environ.pop(env_name, None)
-                    st.session_state.pop(f"ui_{env_name.lower()}", None)
-                    st.session_state.pop(f"ui_{env_name.lower()}_input", None)
-                st.success("✅ Cleared session API keys.")
-                st.rerun()
-
-        st.markdown("---")
-        api_tabs = st.tabs(["OpenAI", "Anthropic", "Google", "Local"])
-
-        with api_tabs[0]:
-            config["api"]["openai"]["base_url"] = st.text_input(
-                "Base URL",
-                value=config["api"]["openai"]["base_url"],
-                key="config_openai_url",
-            )
-            config["api"]["openai"]["argo_user"] = st.text_input(
-                "Argo User (optional)",
-                value=config["api"]["openai"].get("argo_user", ""),
-                key="config_openai_argo_user",
-                help="ANL domain username for Argo requests. If blank, ChemGraph falls back to ARGO_USER env var.",
-            )
-            config["api"]["openai"]["timeout"] = st.number_input(
-                "Timeout (seconds)",
-                min_value=1,
-                max_value=300,
-                value=config["api"]["openai"]["timeout"],
-                key="config_openai_timeout",
-            )
-
-        with api_tabs[1]:
-            config["api"]["anthropic"]["base_url"] = st.text_input(
-                "Base URL",
-                value=config["api"]["anthropic"]["base_url"],
-                key="config_anthropic_url",
-            )
-            config["api"]["anthropic"]["timeout"] = st.number_input(
-                "Timeout (seconds)",
-                min_value=1,
-                max_value=300,
-                value=config["api"]["anthropic"]["timeout"],
-                key="config_anthropic_timeout",
-            )
-
-        with api_tabs[2]:
-            config["api"]["google"]["base_url"] = st.text_input(
-                "Base URL",
-                value=config["api"]["google"]["base_url"],
-                key="config_google_url",
-            )
-            config["api"]["google"]["timeout"] = st.number_input(
-                "Timeout (seconds)",
-                min_value=1,
-                max_value=300,
-                value=config["api"]["google"]["timeout"],
-                key="config_google_timeout",
-            )
-
-        with api_tabs[3]:
-            config["api"]["local"]["base_url"] = st.text_input(
-                "Base URL",
-                value=config["api"]["local"]["base_url"],
-                key="config_local_url",
-            )
-            config["api"]["local"]["timeout"] = st.number_input(
-                "Timeout (seconds)",
-                min_value=1,
-                max_value=300,
-                value=config["api"]["local"]["timeout"],
-                key="config_local_timeout",
-            )
-
-    with tab3:
-        st.subheader("Raw TOML Configuration")
-        st.markdown(
-            """
-        Edit the raw TOML configuration directly. Be careful with syntax!
-        """
-        )
-
-        try:
-            config_text = toml.dumps(config)
-        except Exception as e:
-            st.error(f"Error serializing config: {e}")
-            config_text = ""
-
-        edited_config = st.text_area(
-            "TOML Content", value=config_text, height=400, key="config_raw_toml"
-        )
-
-        if st.button("📝 Update from TOML", key="update_from_toml"):
-            try:
-                new_config = toml.loads(edited_config)
-                st.session_state.config = new_config
-                st.success("✅ Configuration updated from TOML!")
-                st.rerun()
-            except Exception as e:
-                st.error(f"❌ Invalid TOML syntax: {e}")
-
-    # Action buttons
-    st.markdown("---")
-    col1, col2, col3, col4 = st.columns(4)
-
-    with col1:
-        if st.button("💾 Save Configuration", type="primary"):
-            if save_config(config):
-                st.success("✅ Configuration saved to config.toml!")
-            else:
-                st.error("❌ Failed to save configuration")
-
-    with col2:
-        if st.button("🔄 Reload Configuration"):
-            st.session_state.config = load_config()
-            st.success("✅ Configuration reloaded!")
-            st.rerun()
-
-    with col3:
-        if st.button("🗑️ Reset to Defaults"):
-            st.session_state.config = get_default_config()
-            st.success("✅ Configuration reset to defaults!")
-            st.rerun()
-
-    with col4:
-        # Download button for config file
-        try:
-            config_download = toml.dumps(config)
-            st.download_button(
-                "📥 Download TOML",
-                config_download,
-                "config.toml",
-                mime="application/toml",
-            )
-        except Exception as e:
-            st.error(f"Error preparing download: {e}")
-
-    # Configuration preview
-    with st.expander("📊 Configuration Summary", expanded=False):
-        st.write("**Current Configuration:**")
-        st.write(f"- Model: {config['general']['model']}")
-        st.write(f"- Workflow: {config['general']['workflow']}")
-        st.write("- Temperature: 0.0 (optimized for tool calling)")
-        st.write("- Max Tokens: 4000")
-        st.write(
-            f"- Default Calculator: {config['chemistry']['calculators']['default']}"
-        )
-
-        # Environment variables check
-        st.write("**Environment Variables:**")
-        api_keys = {
-            "OPENAI_API_KEY": "OpenAI",
-            "ANTHROPIC_API_KEY": "Anthropic",
-            "GEMINI_API_KEY": "Google",
-            "GROQ_API_KEY": "Groq",
-        }
-
-        for env_var, provider in api_keys.items():
-            if os.getenv(env_var):
-                st.write(f"- {provider}: ✅ Set")
-            else:
-                st.write(f"- {provider}: ❌ Not set")
-
-    # Stop execution here for Config page
+    configuration.render()
     st.stop()
 
-# -----------------------------------------------------------------------------
-# Main Interface (only runs if not on About or Config page)
-# -----------------------------------------------------------------------------
-
-# -----------------------------------------------------------------------------
-# Main title & description
-# -----------------------------------------------------------------------------
-# -----------------------------------------------------------------------------
-# Session-state init and configuration loading (MUST BE FIRST)
-# -----------------------------------------------------------------------------
-if "agent" not in st.session_state:
-    st.session_state.agent = None
-if "conversation_history" not in st.session_state:
-    st.session_state.conversation_history = []
-if "last_config" not in st.session_state:
-    st.session_state.last_config = None
-if "config" not in st.session_state:
-    st.session_state.config = load_config()
-if "last_run_error" not in st.session_state:
-    st.session_state.last_run_error = None
-if "last_run_result" not in st.session_state:
-    st.session_state.last_run_result = None
-if "last_run_query" not in st.session_state:
-    st.session_state.last_run_query = None
-if "ui_notice" not in st.session_state:
-    st.session_state.ui_notice = None
-
-# Get configuration values
-config = st.session_state.config
-selected_model = config["general"]["model"]
-selected_workflow = normalize_workflow_name(config["general"]["workflow"])
-selected_output = config["general"]["output"]
-structured_output = config["general"]["structured"]
-generate_report = config["general"]["report"]
-thread_id = config["general"]["thread"]
-
-# Argo OpenAI-compatible endpoint often returns plain text; disable structured output.
-if selected_model in supported_argo_models and structured_output:
-    structured_output = False
-    st.session_state.ui_notice = (
-        "Structured output is disabled for Argo models to avoid JSON parsing errors."
-    )
-
-# -----------------------------------------------------------------------------
-# Main Interface Header
-# -----------------------------------------------------------------------------
-
-st.title("🧪 ChemGraph")
-
-st.markdown(
-    """
-ChemGraph enables you to perform various **computational chemistry** tasks with
-natural-language queries using AI agents.
-"""
-)
-
-# Quick settings override
-with st.sidebar.expander("🔧 Quick Settings"):
-    st.write("Override settings for this session:")
-
-    # Model override
-    if st.checkbox("Override Model"):
-        model_options = get_model_options(config)
-        selected_model = st.selectbox(
-            "Select Model",
-            model_options,
-            index=(
-                model_options.index(selected_model)
-                if selected_model in model_options
-                else 0
-            ),
-        )
-        quick_custom_model = st.text_input(
-            "Custom model ID (optional)",
-            value="",
-            key="quick_custom_model",
-            help="If set, this overrides the selected model for this session.",
-        ).strip()
-        if quick_custom_model:
-            selected_model = quick_custom_model
-
-    # Thread ID override
-    if st.checkbox("Override Thread ID"):
-        thread_id = st.number_input(
-            "Thread ID", min_value=1, max_value=1000, value=thread_id
-        )
-
-    st.info("💡 To make permanent changes, use the Configuration page.")
-
-selected_base_url = get_base_url_for_model(selected_model, config)
-endpoint_status = check_local_model_endpoint(selected_base_url)
-
-# Reload config button
-if st.sidebar.button("🔄 Reload Config"):
-    st.session_state.config = load_config()
-    st.success("✅ Configuration reloaded!")
-    st.rerun()
-
-# -----------------------------------------------------------------------------
-# Agent status section
-# -----------------------------------------------------------------------------
-st.sidebar.header("🅒🅖 Agent Status")
-
-if st.session_state.agent:
-    st.sidebar.success("✅ Agents Ready")
-    st.sidebar.info(f"🧠 Model: {selected_model}")
-    st.sidebar.info(f"⚙️ Workflow: {selected_workflow}")
-    st.sidebar.info(f"🔗 Thread ID: {thread_id}")
-    st.sidebar.info(f"💬 Messages: {len(st.session_state.conversation_history)}")
-    if endpoint_status["ok"]:
-        st.sidebar.caption(f"LLM endpoint: {endpoint_status['message']}")
-    else:
-        st.sidebar.error(f"LLM endpoint issue: {endpoint_status['message']}")
-    if st.session_state.last_run_error:
-        st.sidebar.error("Last run error (see verbose info).")
-
-    # Add a manual refresh button for troubleshooting
-    if st.sidebar.button("🔄 Refresh Agents"):
-        st.session_state.agent = None  # Force re-initialization
-        st.rerun()
 else:
-    st.sidebar.error("❌ Agents Not Ready")
-    st.sidebar.info("Agents will initialize automatically...")
-    if not endpoint_status["ok"]:
-        st.sidebar.error(f"LLM endpoint issue: {endpoint_status['message']}")
+    from ui._pages import main_interface
 
-# Configuration page link
-st.sidebar.markdown("---")
-st.sidebar.markdown("**⚙️ Configuration**")
-st.sidebar.markdown(
-    "Use the Configuration page to modify settings, API endpoints, and chemistry parameters."
-)
-st.sidebar.markdown("Current config loaded from: `config.toml`")
-
-# -----------------------------------------------------------------------------
-# Helper: check if IR spectrum file has changed within last minute
-# -----------------------------------------------------------------------------
-
-
-def changed_recently(path="ir_spectrum.png", window_seconds=300) -> bool:
-    """
-    Return True if `path` exists and was modified within the last `window_seconds`.
-    """
-    p = Path(resolve_output_path(path))
-    if not p.exists():
-        return False
-
-    mtime = datetime.fromtimestamp(p.stat().st_mtime, tz=timezone.utc)
-    now = datetime.now(timezone.utc)
-    return (now - mtime) <= timedelta(seconds=window_seconds)
-
-
-# -----------------------------------------------------------------------------
-# Helper: extract molecular structure from plain-text message
-# -----------------------------------------------------------------------------
-def find_html_filename(messages: list) -> Optional[str]:
-    """
-    Scan through *messages* in reverse order for the first occurrence of something
-    that looks like an HTML file (e.g. 'report.html' or 'results/2025/plot.html').
-    Returns the matched substring (path or bare filename) or `None` if nothing is found.
-
-    Parameters
-    ----------
-    messages : list
-        List of message objects to search through
-
-    Returns
-    -------
-    str or None
-        HTML filename/path if found, None otherwise
-
-    Examples
-    --------
-    >>> messages = [{"content": "See docs in build/output/index.html"}, {"content": "No HTML"}]
-    >>> find_html_filename(messages)
-    'build/output/index.html'
-
-    >>> find_html_filename([{"content": "No HTML here"}])
-    None
-    """
-    pattern = r"[\w./-]+\.html\b"  # words / dots / slashes up to '.html'
-
-    # Search through messages in reverse order (most recent first)
-    for message in reversed(messages):
-        # Extract content from different message formats
-        raw_content = ""
-        if hasattr(message, "content"):
-            raw_content = getattr(message, "content", "")
-        elif isinstance(message, dict):
-            raw_content = message.get("content", "")
-        elif isinstance(message, str):
-            raw_content = message
-        else:
-            raw_content = str(message)
-        content = normalize_message_content(raw_content)
-
-        # Search for HTML pattern in this message content
-        if content:
-            match = re.search(pattern, content, flags=re.IGNORECASE)
-            if match:
-                return match.group(0)  # Return immediately when found
-
-    return None  # No HTML filename found in any message
-
-
-def extract_xyz_from_report_html(html_content: str):
-    """Extract atomic numbers and positions from a ChemGraph HTML report.
-
-    The report embeds XYZ data as a base64-encoded string passed to ``atob()``.
-    This function decodes that string, parses the XYZ atom lines, and converts
-    element symbols to atomic numbers using ASE's ``chemical_symbols`` table.
-
-    Parameters
-    ----------
-    html_content : str
-        The full HTML string of a ChemGraph report.
-
-    Returns
-    -------
-    dict or None
-        ``{"atomic_numbers": [...], "positions": [...]}`` on success, else
-        ``None``.
-    """
-    import base64 as _b64
-
-    match = re.search(r'atob\(["\']([A-Za-z0-9+/=]+)["\']\)', html_content)
-    if not match:
-        return None
-
-    try:
-        xyz_text = _b64.b64decode(match.group(1)).decode("utf-8")
-    except Exception:
-        return None
-
-    lines = xyz_text.strip().splitlines()
-    if len(lines) < 3:
-        return None
-
-    try:
-        num_atoms = int(lines[0].strip())
-    except ValueError:
-        return None
-
-    # Build a reverse lookup: symbol -> atomic number
-    sym_to_num = {s: i for i, s in enumerate(chemical_symbols)}
-
-    atomic_numbers = []
-    positions = []
-    for line in lines[2 : 2 + num_atoms]:
-        parts = line.strip().split()
-        if len(parts) < 4:
-            continue
-        symbol = parts[0]
-        anum = sym_to_num.get(symbol)
-        if anum is None:
-            return None  # unknown element – bail out
-        try:
-            pos = [float(parts[1]), float(parts[2]), float(parts[3])]
-        except ValueError:
-            return None
-        atomic_numbers.append(anum)
-        positions.append(pos)
-
-    if len(atomic_numbers) != num_atoms:
-        return None
-
-    return {"atomic_numbers": atomic_numbers, "positions": positions}
-
-
-def strip_viewer_from_report_html(html_content: str) -> str:
-    """Remove the NGL 3D-viewer section from a ChemGraph HTML report.
-
-    Strips the ``<div id="viewer">`` element, the NGL ``<script src>`` tag,
-    and the inline ``<script>`` block that initialises the NGL stage / XYZ
-    parsing so that only the calculation-results and simulation-details
-    sections remain.  The resulting HTML can be safely rendered inside
-    Streamlit's ``st.components.v1.html()`` iframe.
-
-    Parameters
-    ----------
-    html_content : str
-        Full HTML string of a ChemGraph report.
-
-    Returns
-    -------
-    str
-        The HTML with viewer-related elements removed.
-    """
-    # 1. Remove the external NGL script tag
-    html_content = re.sub(
-        r'<script\s+src="[^"]*ngl[^"]*"[^>]*>\s*</script>\s*',
-        "",
-        html_content,
-        flags=re.IGNORECASE,
-    )
-
-    # 2. Remove the <div id="viewer"></div>
-    html_content = re.sub(
-        r'<div\s+id="viewer"[^>]*>\s*</div>\s*',
-        "",
-        html_content,
-    )
-
-    # 3. Remove the inline script block that creates the NGL Stage, xyzToPDB
-    #    function, and loads the molecule.  This block starts with
-    #    ``const stage = new NGL.Stage`` and ends at ``</script>``.
-    html_content = re.sub(
-        r"<script>\s*function toggleSection.*?const stage = new NGL\.Stage.*?</script>",
-        # Keep the toggleSection / toggleSubSection JS (needed for
-        # collapsible sections) but drop everything from the NGL stage
-        # onwards.  Easier to just re-inject the toggle helpers.
-        """<script>
-        function toggleSection(sectionId) {
-            const content = document.getElementById(sectionId);
-            const header = content.previousElementSibling;
-            content.classList.toggle('collapsed');
-            header.classList.toggle('collapsed');
-        }
-        function toggleSubSection(sectionId) {
-            const content = document.getElementById(sectionId);
-            const header = content.previousElementSibling;
-            if (content.style.display === 'none') {
-                content.style.display = 'block';
-                header.classList.remove('collapsed');
-            } else {
-                content.style.display = 'none';
-                header.classList.add('collapsed');
-            }
-        }
-        document.addEventListener('DOMContentLoaded', function() {
-            const subSections = document.querySelectorAll('.sub-section-content');
-            subSections.forEach(section => { section.style.display = 'block'; });
-        });
-        </script>""",
-        html_content,
-        flags=re.DOTALL,
-    )
-
-    # 4. Remove the heading that says "XYZ Molecule Viewer" since the 3D
-    #    viewer is now rendered natively by Streamlit above the report.
-    html_content = re.sub(
-        r"<h1>\s*XYZ Molecule Viewer\s*</h1>\s*",
-        "",
-        html_content,
-    )
-
-    return html_content
-
-
-def extract_molecular_structure(message_content: str):
-    """Return dict with keys atomic_numbers, positions if embedded in message."""
-    if not message_content:
-        return None
-
-    # First try to parse as JSON (for structured output)
-    try:
-        # Check if the content is JSON with structure data
-        if message_content.strip().startswith("{") and message_content.strip().endswith(
-            "}"
-        ):
-            json_data = json.loads(message_content)
-
-            # Look for structure data in various JSON formats
-            structure_data = None
-            if "answer" in json_data:
-                structure_data = json_data["answer"]
-            elif "numbers" in json_data and "positions" in json_data:
-                structure_data = json_data
-            elif "atomic_numbers" in json_data and "positions" in json_data:
-                structure_data = json_data
-
-            if (
-                structure_data
-                and "numbers" in structure_data
-                and "positions" in structure_data
-            ):
-                return {
-                    "atomic_numbers": structure_data["numbers"],
-                    "positions": structure_data["positions"],
-                }
-            elif (
-                structure_data
-                and "atomic_numbers" in structure_data
-                and "positions" in structure_data
-            ):
-                return {
-                    "atomic_numbers": structure_data["atomic_numbers"],
-                    "positions": structure_data["positions"],
-                }
-    except (json.JSONDecodeError, KeyError):
-        pass
-
-    # Then try to parse plain text format (original method)
-    lines = message_content.splitlines()
-    atomic_numbers, positions = None, None
-
-    for i, line in enumerate(lines):
-        if "Atomic Numbers" in line:
-            try:
-                numbers_str = line.split(":")[1].strip()
-                atomic_numbers = ast.literal_eval(numbers_str)
-            except Exception:
-                pass
-        elif "Positions" in line:
-            positions = []
-            for sub in lines[i + 1 :]:
-                sub = sub.strip()
-                if sub.startswith("- [") and sub.endswith("]"):
-                    try:
-                        positions.append(ast.literal_eval(sub[2:]))
-                    except Exception:
-                        pass
-                elif not sub.startswith("-") and positions:
-                    break
-
-    if (
-        isinstance(atomic_numbers, list)
-        and isinstance(positions, list)
-        and len(atomic_numbers) == len(positions)
-    ):
-        return {"atomic_numbers": atomic_numbers, "positions": positions}
-
-    return None
-
-
-# Helper: extract messages from result object
-def extract_messages_from_result(result):
-    """Extract messages from result object, handling different formats."""
-    if isinstance(result, list):
-        return result  # Already a list of messages
-    elif isinstance(result, dict) and "messages" in result:
-        messages = result["messages"]
-
-        # For multi-agent workflows, also extract messages from worker_channel
-        if "worker_channel" in result:
-            worker_channel = result["worker_channel"]
-            # Flatten all worker messages into the main messages list
-            for worker_id, worker_messages in worker_channel.items():
-                if isinstance(worker_messages, list):
-                    messages.extend(worker_messages)
-
-        return messages
-    else:
-        return [result]  # Treat as single message
-
-
-def normalize_message_content(content: Any) -> str:
-    """Convert varying message content payloads (str/list/dict) into plain text."""
-    if content is None:
-        return ""
-    if isinstance(content, str):
-        return content
-    if isinstance(content, list):
-        parts: list[str] = []
-        for item in content:
-            if isinstance(item, str):
-                parts.append(item)
-            elif isinstance(item, dict):
-                text = item.get("text")
-                if isinstance(text, str):
-                    parts.append(text)
-                else:
-                    parts.append(str(item))
-            else:
-                parts.append(str(item))
-        return "\n".join(p for p in parts if p)
-    if isinstance(content, dict):
-        text = content.get("text")
-        if isinstance(text, str):
-            return text
-        return str(content)
-    return str(content)
-
-
-# Helper: find structure data in messages
-def find_structure_in_messages(messages):
-    """Look through all messages to find structure data."""
-    for message in messages:
-        if hasattr(message, "content") or isinstance(message, dict):
-            raw_content = (
-                getattr(message, "content", "")
-                if hasattr(message, "content")
-                else message.get("content", "")
-            )
-            content = normalize_message_content(raw_content)
-            structure = extract_molecular_structure(content)
-            if structure:
-                return structure
-    return None
-
-
-def has_structure_signal(
-    messages, query_text: str = "", final_answer: str = ""
-) -> bool:
-    """Return True when current interaction appears to include structure artifacts."""
-    structure_tools = {
-        "smiles_to_coordinate_file",
-        "run_ase",
-        "file_to_atomsdata",
-        "save_atomsdata_to_file",
-    }
-    structure_markers = (
-        ".xyz",
-        "final_structure",
-        "atomic_numbers",
-        "positions",
-        "coordinate_file",
-    )
-
-    for message in messages:
-        name = getattr(message, "name", None)
-        content = getattr(message, "content", "")
-
-        if isinstance(message, dict):
-            name = message.get("name", name)
-            content = message.get("content", content)
-
-        if name in structure_tools:
-            return True
-
-        if isinstance(content, str):
-            lowered = content.lower()
-            if any(marker in lowered for marker in structure_markers):
-                return True
-
-    combined_text = f"{query_text}\n{final_answer}".lower()
-    keyword_markers = (
-        "geometry",
-        "optimiz",
-        "structure",
-        "coordinates",
-        "xyz",
-    )
-    return any(marker in combined_text for marker in keyword_markers)
-
-
-def is_infrared_requested(messages):
-    """Look through all messages to find infrared data."""
-    for message in messages:
-        # Handle different message formats (string/list/dict payloads)
-        raw_content = ""
-        if hasattr(message, "content"):
-            raw_content = getattr(message, "content", "")
-        elif isinstance(message, dict):
-            raw_content = message.get("content", "")
-        elif isinstance(message, str):
-            raw_content = message
-        else:
-            raw_content = str(message)
-
-        content = normalize_message_content(raw_content)
-        lowered = content.lower()
-        if content and (("infrared" in lowered) or re.search(r"\bir\b", lowered)):
-            return True
-    return False
-
-
-# Streamlit-specific wrapper for ASE functions
-def create_ase_atoms_with_streamlit_error(atomic_numbers, positions):
-    """Wrapper for create_ase_atoms that displays errors in Streamlit."""
-    atoms = create_ase_atoms(atomic_numbers, positions)
-    if atoms is None:
-        st.error("Error creating ASE Atoms object")
-    return atoms
-
-
-# -----------------------------------------------------------------------------
-# Display 3-D (or fallback) molecular structure
-# -----------------------------------------------------------------------------
-def display_molecular_structure(atomic_numbers, positions, title="Structure"):
-    try:
-        atoms = create_ase_atoms_with_streamlit_error(atomic_numbers, positions)
-        if atoms is None:
-            return False
-
-        xyz_string = create_xyz_string(atomic_numbers, positions)
-        if xyz_string is None:
-            return False
-
-        st.subheader(f"🧬 {title}")
-        col1, col2 = st.columns([2, 1])
-
-        # 3-D panel ------------------------------------------------------------
-        with col1:
-            if STMOL_AVAILABLE:
-                style_options = ["ball_and_stick", "stick", "sphere", "wireframe"]
-                selected_style = st.selectbox(
-                    "Visualization Style", style_options, key=f"style_{uuid4().hex}"
-                )
-
-                # Create the 3D visualization using stmol directly
-                try:
-                    import py3Dmol
-
-                    # Create py3Dmol viewer
-                    view = py3Dmol.view(width=500, height=400)
-                    view.addModel(xyz_string, "xyz")
-
-                    if selected_style == "ball_and_stick":
-                        view.setStyle({"stick": {}, "sphere": {"scale": 0.3}})
-                    elif selected_style == "stick":
-                        view.setStyle({"stick": {}})
-                    elif selected_style == "sphere":
-                        view.setStyle({"sphere": {}})
-                    elif selected_style == "wireframe":
-                        view.setStyle({"line": {}})
-                    else:
-                        view.setStyle({"stick": {}, "sphere": {"scale": 0.3}})
-
-                    view.zoomTo()
-
-                    # Use stmol.showmol with the py3Dmol view object
-                    stmol.showmol(view, height=400, width=500)
-
-                except Exception as viz_error:
-                    st.error(f"3D visualization error: {viz_error}")
-                    st.info("Falling back to table view...")
-                    # Show fallback table
-                    data = []
-                    for idx, (num, pos) in enumerate(zip(atomic_numbers, positions), 1):
-                        sym = (
-                            chemical_symbols[num]
-                            if num < len(chemical_symbols)
-                            else f"X{num}"
-                        )
-                        data.append(
-                            {
-                                "Atom": idx,
-                                "Element": sym,
-                                "X": f"{pos[0]:.4f}",
-                                "Y": f"{pos[1]:.4f}",
-                                "Z": f"{pos[2]:.4f}",
-                            }
-                        )
-                    st.dataframe(
-                        pd.DataFrame(data), height=350, use_container_width=True
-                    )
-            else:
-                st.info("3-D viewer unavailable; showing raw XYZ and table.")
-
-                # Show XYZ content
-                with st.expander("📄 XYZ Format", expanded=True):
-                    st.code(xyz_string, language="text")
-
-                # Show structure table
-                data = []
-                for idx, (num, pos) in enumerate(zip(atomic_numbers, positions), 1):
-                    sym = (
-                        chemical_symbols[num]
-                        if num < len(chemical_symbols)
-                        else f"X{num}"
-                    )
-                    data.append(
-                        {
-                            "Atom": idx,
-                            "Element": sym,
-                            "X": f"{pos[0]:.4f}",
-                            "Y": f"{pos[1]:.4f}",
-                            "Z": f"{pos[2]:.4f}",
-                        }
-                    )
-                st.dataframe(pd.DataFrame(data), height=350, use_container_width=True)
-
-        # Info panel -----------------------------------------------------------
-        with col2:
-            st.markdown("**Structure Information**")
-            st.write(f"- **Atoms:** {len(atoms)}")
-            st.write(f"- **Formula:** {atoms.get_chemical_formula()}")
-
-            # Composition
-            composition = {}
-            for atom in atoms:
-                composition[atom.symbol] = composition.get(atom.symbol, 0) + 1
-            st.write("**Composition:**")
-            for elem, count in sorted(composition.items()):
-                st.write(f"  • {elem}: {count}")
-
-            # Total mass
-            try:
-                total_mass = atoms.get_masses().sum()
-                st.write(f"**Total Mass:** {total_mass:.2f} amu")
-            except Exception:
-                st.write("**Total Mass:** Not available")
-
-            # Center of mass
-            try:
-                com = atoms.get_center_of_mass()
-                st.write("**Center of Mass:**")
-                st.write(f"  [{com[0]:.3f}, {com[1]:.3f}, {com[2]:.3f}] Å")
-            except Exception:
-                st.write("**Center of Mass:** Not available")
-
-            # Additional properties
-            with st.expander("🔬 Additional Properties"):
-                try:
-                    pos = atoms.positions
-                    com = atoms.get_center_of_mass()
-                    distances = np.linalg.norm(pos - com, axis=1)
-                    st.write(f"**Max distance from COM:** {distances.max():.3f} Å")
-                    st.write(f"**Min distance from COM:** {distances.min():.3f} Å")
-
-                    cell = atoms.get_cell()
-                    if np.any(cell.lengths()):  # any non-zero → periodic
-                        st.write(f"**Cell lengths:** {cell.lengths()}")
-                        st.write(f"**Cell angles:** {cell.angles()}")
-                    else:
-                        st.write("**Cell:** non-periodic")
-                except Exception as prop_error:
-                    st.write(f"Error calculating properties: {prop_error}")
-
-            # Downloads
-            st.write("**Download:**")
-            st.download_button(
-                "📄 XYZ File",
-                xyz_string,
-                f"{title.lower().replace(' ', '_')}.xyz",
-                mime="chemical/x-xyz",
-                key=f"xyz_download_{uuid4().hex}",
-            )
-
-            structure_json = json.dumps(
-                {
-                    "atomic_numbers": atomic_numbers,
-                    "positions": positions,
-                    "formula": atoms.get_chemical_formula(),
-                    "symbols": atoms.get_chemical_symbols(),
-                },
-                indent=2,
-            )
-            st.download_button(
-                "📋 JSON Data",
-                structure_json,
-                f"{title.lower().replace(' ', '_')}.json",
-                mime="application/json",
-                key=f"json_download_{uuid4().hex}",
-            )
-
-        return True
-    except Exception as exc:
-        st.error(f"Error displaying structure: {exc}")
-    return False
-
-
-def visualize_trajectory(traj):
-    """Create an animated 3D visualization of a trajectory.
-
-    Args:
-        traj: ASE Trajectory object
-
-    Returns:
-        view: py3Dmol view object with animated trajectory
-    """
-    # Convert all frames to a single multi-model XYZ string
-    import py3Dmol
-
-    xyz_frames = []
-    for i, atoms in enumerate(traj):
-        symbols = atoms.get_chemical_symbols()
-        pos = atoms.get_positions()  # Å
-        lines = [str(len(symbols)), f"Frame {i}"]
-        lines += [f"{s} {x:.6f} {y:.6f} {z:.6f}" for s, (x, y, z) in zip(symbols, pos)]
-        xyz_frames.append("\n".join(lines))
-    xyz_str = "\n".join(xyz_frames)
-
-    # Initialize viewer and add frames
-    view = py3Dmol.view(width=500, height=400)
-    view.addModelsAsFrames(xyz_str, "xyz")  # load all frames at once
-
-    # Style & camera
-    view.setViewStyle({"style": "outline", "width": 0.05})
-    view.setStyle({"stick": {}, "sphere": {"scale": 0.25}})
-    view.zoomTo()
-
-    # Animate (interval in ms)
-    view.animate({"loop": "Forward", "interval": 100})
-
-    return view
-
-
-def find_latest_xyz_file() -> Optional[str]:
-    """Find the most recently modified .xyz file in the log dir or cwd."""
-    search_dirs = []
-    log_dir = os.environ.get("CHEMGRAPH_LOG_DIR")
-    if log_dir:
-        search_dirs.append(log_dir)
-    search_dirs.append(os.getcwd())
-
-    latest_path = None
-    latest_mtime = -1.0
-    for base in search_dirs:
-        if not base or not os.path.isdir(base):
-            continue
-        for path in Path(base).rglob("*.xyz"):
-            try:
-                mtime = path.stat().st_mtime
-            except OSError:
-                continue
-            if mtime > latest_mtime:
-                latest_mtime = mtime
-                latest_path = str(path)
-    return latest_path
-
-
-def find_latest_xyz_file_in_dir(directory: str) -> Optional[str]:
-    """Find the most recently modified .xyz file under a specific directory."""
-    if not directory or not os.path.isdir(directory):
-        return None
-    latest_path = None
-    latest_mtime = -1.0
-    for path in Path(directory).rglob("*.xyz"):
-        try:
-            mtime = path.stat().st_mtime
-        except OSError:
-            continue
-        if mtime > latest_mtime:
-            latest_mtime = mtime
-            latest_path = str(path)
-    return latest_path
-
-
-def extract_log_dir_from_messages(messages) -> Optional[str]:
-    """Extract a directory path from message content that references an output file."""
-    if not messages:
-        return None
-    patterns = [
-        r"(/[^\s'\"`]+?\.json)",
-        r"(/[^\s'\"`]+?\.xyz)",
-        r"(/[^\s'\"`]+?\.html)",
-        r"(/[^\s'\"`]+?\.csv)",
-    ]
-
-    def _scan_value(value):
-        if isinstance(value, str):
-            for pattern in patterns:
-                match = re.search(pattern, value)
-                if match:
-                    path = match.group(1)
-                    if os.path.isabs(path):
-                        return str(Path(path).parent)
-        elif isinstance(value, dict):
-            for v in value.values():
-                found = _scan_value(v)
-                if found:
-                    return found
-        elif isinstance(value, list):
-            for v in value:
-                found = _scan_value(v)
-                if found:
-                    return found
-        return None
-
-    for message in reversed(messages):
-        content = ""
-        if hasattr(message, "content"):
-            content = getattr(message, "content", "")
-        elif isinstance(message, dict):
-            content = message.get("content", "")
-        elif isinstance(message, str):
-            content = message
-        else:
-            content = str(message)
-        if not content:
-            continue
-        found = _scan_value(content)
-        if found:
-            return found
-
-        # Also scan structured tool outputs if present
-        if hasattr(message, "additional_kwargs"):
-            found = _scan_value(message.additional_kwargs)
-            if found:
-                return found
-        if isinstance(message, dict):
-            found = _scan_value(message)
-            if found:
-                return found
-    return None
-
-
-# Function for IR spectrum rendering
-
-
-# -----------------------------------------------------------------------------
-# Agent initializer (cached)
-# -----------------------------------------------------------------------------
-@st.cache_resource
-def initialize_agent(
-    model_name,
-    workflow_type,
-    structured_output,
-    return_option,
-    generate_report,
-    recursion_limit,
-    base_url,
-    argo_user,
-):
-    try:
-        from chemgraph.agent.llm_agent import ChemGraph
-
-        return ChemGraph(
-            model_name=model_name,
-            workflow_type=workflow_type,
-            base_url=base_url,
-            argo_user=argo_user,
-            structured_output=structured_output,
-            generate_report=generate_report,
-            return_option=return_option,
-            recursion_limit=recursion_limit,
-        )
-    except Exception as exc:
-        st.error(f"Failed to initialize agent: {exc}")
-        return None
-
-
-# -----------------------------------------------------------------------------
-# Auto-initialize agent when configuration changes
-# -----------------------------------------------------------------------------
-current_config = (
-    selected_model,
-    selected_workflow,
-    structured_output,
-    selected_output,
-    generate_report,
-    config["general"]["recursion_limit"],
-    selected_base_url,
-    get_argo_user_from_nested_config(config),
-)
-
-if st.session_state.agent is None or st.session_state.last_config != current_config:
-    with st.spinner("🚀 Initializing ChemGraph agents..."):
-        st.session_state.agent = initialize_agent(
-            selected_model,
-            selected_workflow,
-            structured_output,
-            selected_output,
-            generate_report,
-            config["general"]["recursion_limit"],
-            selected_base_url,
-            get_argo_user_from_nested_config(config),
-        )
-        st.session_state.last_config = current_config
-
-
-# -----------------------------------------------------------------------------
-# Main chat interface
-# -----------------------------------------------------------------------------
-
-# Conversation history display
-if st.session_state.conversation_history:
-    st.subheader("🗨️ Conversation History")
-    for idx, entry in enumerate(st.session_state.conversation_history, 1):
-        # User bubble
-        st.markdown(
-            f"""
-<div style="background:#e3f2fd;padding:15px;border-radius:15px;margin:10px 0 0 50px;border:1px solid #2196f3;color:#000000;">
-  <b style="color:#1976d2;">👤 You:</b><br><span style="color:#333333;">{html_mod.escape(entry["query"])}</span>
-</div>""",
-            unsafe_allow_html=True,
-        )
-
-        # Extract messages from the result
-        messages = extract_messages_from_result(entry["result"])
-
-        # Find the final AI response for display
-        final_answer = ""
-        for message in reversed(messages):
-            # Handle different message formats
-            if hasattr(message, "content") and hasattr(message, "type"):
-                # LangChain message object
-                content = normalize_message_content(message.content).strip()
-                if message.type == "ai" and content:
-                    # Skip if it's just JSON structure data
-                    if not (
-                        content.startswith("{")
-                        and content.endswith("}")
-                        and "numbers" in content
-                    ):
-                        final_answer = content
-                        break
-            elif isinstance(message, dict):
-                # Dictionary message format
-                content = normalize_message_content(message.get("content", "")).strip()
-                if message.get("type") == "ai" and content:
-                    if not (
-                        content.startswith("{")
-                        and content.endswith("}")
-                        and "numbers" in content
-                    ):
-                        final_answer = content
-                        break
-            elif hasattr(message, "content"):
-                # Generic message object with content
-                content = normalize_message_content(
-                    getattr(message, "content", "")
-                ).strip()
-                if content and not (
-                    content.startswith("{")
-                    and content.endswith("}")
-                    and "numbers" in content
-                ):
-                    final_answer = content
-                    break
-
-        # Display the AI response
-        if final_answer:
-            st.markdown(
-                f"""
-<div style="background:#f1f8e9;padding:15px;border-radius:15px;margin:10px 50px 0 0;border:1px solid #4caf50;color:#000000;">
-  <b style="color:#388e3c;">🅒🅖 ChemGraph:</b><br><span style="color:#333333;">{html_mod.escape(final_answer).replace(chr(10), "<br>")}</span>
-</div>""",
-                unsafe_allow_html=True,
-            )
-
-        # Look for structure data across all messages.
-        # Compute html_filename early so we can skip the .xyz file
-        # fallback when an HTML report will render the same structure.
-        html_filename = find_html_filename(messages)
-
-        structure = find_structure_in_messages(messages)
-        if structure:
-            display_molecular_structure(
-                structure["atomic_numbers"],
-                structure["positions"],
-                title=f"Molecular Structure (Query {idx})",
-            )
-        else:
-            # Also check the final answer text for structure data
-            structure_from_text = extract_molecular_structure(final_answer)
-            if structure_from_text:
-                display_molecular_structure(
-                    structure_from_text["atomic_numbers"],
-                    structure_from_text["positions"],
-                    title=f"Structure from Response {idx}",
-                )
-            elif not html_filename:
-                # Fall back to .xyz file search only when there is no
-                # HTML report that already embeds the same structure.
-                if has_structure_signal(messages, entry.get("query", ""), final_answer):
-                    log_dir = extract_log_dir_from_messages(messages)
-                    if log_dir and os.path.isdir(log_dir):
-                        latest_xyz = find_latest_xyz_file_in_dir(log_dir)
-                        if latest_xyz:
-                            try:
-                                atoms = ase_read(latest_xyz)
-                                display_molecular_structure(
-                                    atoms.get_atomic_numbers().tolist(),
-                                    atoms.get_positions().tolist(),
-                                    title=f"Structure from {Path(latest_xyz).name}",
-                                )
-                            except Exception as exc:
-                                st.warning(f"Failed to load XYZ structure: {exc}")
-        if html_filename:
-            with st.expander("📊 Report", expanded=False):
-                try:
-                    resolved_html = resolve_output_path(html_filename)
-                    with open(resolved_html, "r", encoding="utf-8") as f:
-                        html_content = f.read()
-
-                    # Render the 3D molecule using py3Dmol/stmol
-                    # (NGL.js cannot load inside Streamlit's sandboxed iframe)
-                    report_structure = extract_xyz_from_report_html(html_content)
-                    if report_structure:
-                        display_molecular_structure(
-                            report_structure["atomic_numbers"],
-                            report_structure["positions"],
-                            title="Molecular Structure",
-                        )
-
-                    # Strip the NGL viewer from the HTML and render the
-                    # remaining calculation results / simulation details.
-                    cleaned_html = strip_viewer_from_report_html(html_content)
-                    st.components.v1.html(
-                        cleaned_html, height=600, scrolling=True
-                    )
-                except FileNotFoundError:
-                    st.warning(f"HTML file '{html_filename}' not found")
-                except Exception as e:
-                    st.error(f"Error displaying HTML: {e}")
-
-        # Check for embedded HTML plots/snippets in all messages
-
-        if is_infrared_requested(messages):
-            if changed_recently():
-                with st.expander("🔍 IR Spectrum", expanded=True):
-                    col1, col2 = st.columns(2, border=True)
-
-                    with col1:
-                        ir_path = resolve_output_path("ir_spectrum.png")
-                        if os.path.exists(ir_path):
-                            st.image(ir_path)
-                        else:
-                            st.warning("IR spectrum plot not found.")
-                    with col2:
-                        freq_path = resolve_output_path("frequencies.csv")
-                        if not os.path.exists(freq_path):
-                            st.warning("Frequencies file not found.")
-                        else:
-                            df = pd.read_csv(
-                                freq_path,
-                                index_col=False,
-                                names=["filename", "frequency"],
-                            ).iloc[
-                                6:
-                            ]  # remove the first 6 translation/rotation modes
-
-                            if not df.empty:
-                                # Create a dropdown menu for frequency selection
-                                st.write("**Select a frequency to visualize:**")
-                                freq_options = {
-                                    f"{float(row['frequency'].strip('i')):.2f} cm⁻¹": i
-                                    for i, row in df.iterrows()
-                                }
-                                selected_freq = st.selectbox(
-                                    "Frequency",
-                                    list(freq_options.keys()),
-                                    index=0,
-                                    key=f"ir_frequency_select_{idx}",
-                                )
-                                traj_file = df.loc[freq_options[selected_freq]][
-                                    "filename"
-                                ]
-                                traj_path = resolve_output_path(traj_file)
-                                if not os.path.exists(traj_path):
-                                    st.warning(
-                                        f"Trajectory file '{traj_file}' not found."
-                                    )
-                                elif not STMOL_AVAILABLE:
-                                    st.info(
-                                        "3D viewer not available; install stmol to animate trajectories."
-                                    )
-                                else:
-                                    from ase.io.trajectory import Trajectory
-
-                                    traj = Trajectory(traj_path)
-                                    view = visualize_trajectory(traj)
-                                    # _, center_col, _ = st.columns(3)
-                                    # with center_col:
-                                    view.zoomTo()
-                                    stmol.showmol(view, height=400, width=500)
-                            else:
-                                st.warning("No vibrational frequencies found.")
-
-            else:
-                st.warning("IR spectrum not found.")
-
-        # Optional debug information
-        with st.expander(f"🔍 Verbose Info (Query {idx})", expanded=False):
-            st.write(f"**Number of messages:** {len(messages)}")
-            st.write(f"**Structure found:** {'Yes' if structure else 'No'}")
-            if st.session_state.last_run_query == entry.get("query"):
-                if st.session_state.last_run_error:
-                    st.write("**Last run error:**")
-                    st.code(str(st.session_state.last_run_error))
-                if st.session_state.last_run_result is not None:
-                    st.write("**Raw result (repr):**")
-                    st.code(repr(st.session_state.last_run_result))
-
-            # Show message types and content summaries
-            for i, msg in enumerate(messages):
-                if hasattr(msg, "type"):
-                    msg_type = msg.type
-                    content = normalize_message_content(msg.content)
-                elif isinstance(msg, dict):
-                    msg_type = msg.get("type", "unknown")
-                    content = normalize_message_content(msg.get("content", ""))
-                else:
-                    msg_type = type(msg).__name__
-                    content = normalize_message_content(
-                        getattr(msg, "content", str(msg))
-                    )
-                content_preview = (
-                    (content[:100] + "...") if len(content) > 100 else content
-                )
-
-                st.write(f"  **Message {i+1}:** `{msg_type}` - {content_preview}")
-
-        st.markdown("---")
-
-# -----------------------------------------------------------------------------
-# New query input
-# -----------------------------------------------------------------------------
-
-with st.expander("💡 Example Queries"):
-    st.markdown("**Based on your current configuration:**")
-    st.markdown(f"- Model: {selected_model}")
-    st.markdown(
-        f"- Default Calculator: {config['chemistry']['calculators']['default']}"
-    )
-    st.markdown("- Temperature: 0.0 (optimized for tool calling)")
-
-    examples = [
-        "What is the SMILES string for caffeine?",
-        f"Optimize the geometry of water molecule using {config['chemistry']['calculators']['default']}",
-        "Calculate the infrared spectrum of methanol with xtb calculator",
-        "What is the reaction enthalpy of methane combustion",
-    ]
-    for ex in examples:
-        if st.button(ex, key=f"ex_{ex}"):
-            # Set the example text directly in the text area state
-            st.session_state.query_input = ex
-            st.rerun()
-
-# Initialize query input if not exists
-if "query_input" not in st.session_state:
-    st.session_state.query_input = ""
-
-query = st.text_area(
-    "Enter your computational chemistry query:",
-    value=st.session_state.query_input,
-    height=100,
-    key="query_text_area",  # Different key to avoid conflicts
-)
-
-# Update session state with current text area value
-if query != st.session_state.query_input:
-    st.session_state.query_input = query
-
-col_send, col_clear, col_refresh = st.columns([2, 1, 1])
-
-send = col_send.button("🚀 Send", type="primary", use_container_width=True)
-if col_clear.button("🗑️ Clear Chat", use_container_width=True):
-    st.session_state.conversation_history.clear()
-    # Clear the query input
-    st.session_state.query_input = ""
-    st.rerun()
-if col_refresh.button("🔄 Refresh", use_container_width=True):
-    st.rerun()
-
-# -----------------------------------------------------------------------------
-# Submit query
-# -----------------------------------------------------------------------------
-if send:
-    if not endpoint_status["ok"]:
-        msg = (
-            f"Cannot reach local model endpoint `{selected_base_url}`. "
-            f"{endpoint_status['message']}"
-        )
-        st.session_state.last_run_error = RuntimeError(msg)
-        st.error(msg)
-    elif not st.session_state.agent:
-        st.error("❌ Agent not ready. Please check configuration and try again.")
-        if st.button("🔄 Try Again"):
-            st.rerun()
-    elif not query.strip():
-        st.warning("Please enter a question.")
-    else:
-        with st.spinner("ChemGraph agents working...", show_time=True):
-            try:
-                cfg = {"configurable": {"thread_id": thread_id}}
-                st.session_state.last_run_query = query.strip()
-                st.session_state.last_run_error = None
-                st.session_state.last_run_result = None
-                result = run_async_callable(
-                    lambda: st.session_state.agent.run(query.strip(), config=cfg)
-                )
-                st.session_state.last_run_result = result
-                st.session_state.conversation_history.append(
-                    {"query": query.strip(), "result": result, "thread_id": thread_id}
-                )
-                # Clear the input after successful processing
-                st.session_state.query_input = ""
-                st.success("✅ Done!")
-                st.rerun()
-            except Exception as exc:
-                st.session_state.last_run_error = exc
-                st.error(f"Processing error: {exc}")
-
-# -----------------------------------------------------------------------------
-# Footer
-# -----------------------------------------------------------------------------
-st.markdown("---")
-st.markdown(
-    """
-### Quick Help
-
-**Main Features:** Molecular optimization, vibrational frequencies, SMILES ↔ structure conversions, 3D visualization
-
-📖 For detailed information, documentation, and links to research papers, visit the **About ChemGraph** page.
-"""
-)
-
-if st.session_state.ui_notice:
-    st.info(st.session_state.ui_notice)
+    main_interface.render()
